Add @chanzuckerberg/axe-storybook as a local dependency of the demo app
For easier development. Now we don't have to mess around with "yarn link".

Unfortunately, there's a catch. Everything would be perfect if we could use "yarn add link:../../../" (since it symlinks the dependency, and updates are reflected without us having to do anything). But, there's https://github.com/yarnpkg/yarn/issues/8290...

There's a bug where using "yarn add link:" is not copying over the bin executables.

For now, we'll just use the local dependency. If we update the package, we'll need to run "yarn upgrade @chanzuckerberg/axe-storybook" before we can actually see the changes reflected in the demo app.