Only return the Story properties we need from the browser through playwright
This is necessary because Playwright's `page.evaluate` requires return values to be JSON
serializable, so we need to make sure there are no non-serializable things in this object.
There's no telling what Storybook addons people are using, and whether their parameters are
serializable or not.

See https://github.com/chanzuckerberg/axe-storybook-testing/issues/44 for a bug caused by this.