Re-arrange CI workflow so build happens before lint
bin/axe-storybook.js references a file in build/, so the build needs to happen first. Otherwise we get a lint error.

I have left the lint suppression in (even though it is not strictly necessary now), so that there are no errors when someone first clones the repo but has not built the project, yet.